Добавить кнопку для старта игры
